pyproject.toml
[project]
name = "a2a-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.23",
    "crewai>=1.6.1",
    "google-adk>=1.25.1",
    "langchain>=1.2.10",
    "langchain-core>=1.2.13",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.8",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.2.0",
    "starlette>=0.52.1",
    "uvicorn>=0.41.0",
]
